+++ 2.Logic-Layer/PresupuestosServices/src/dao/DAOUserRoll.java
@@ -1,5 +1,6 @@
 package dao;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.sql2o.Connection;
@@ -26,7 +27,7 @@ public class DAOUserRoll {
 		}
 	}
 	
-	public static boolean insert(long idUser, long rol) {
+	public static boolean insert(BigInteger idUser, BigInteger rol) {
 		initDriver();
 		
 		try (Connection connection = new Sql2o(dataBase,dataBaseUser,dataBasePass).beginTransaction()){
@@ -45,7 +46,7 @@ public class DAOUserRoll {
 		}
 	}
 	
-	public static boolean deleteUserRoll(long idUsuario) {
+	public static boolean deleteUserRoll(BigInteger idUsuario) {
 		initDriver();
 		try (Connection connection = new Sql2o(dataBase,dataBaseUser,dataBasePass).beginTransaction()){
 			String query="delete from user_role where user_role.idUser = :idUser";
